Backened/core/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.3.1
Pillow==11.3.0 